% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot Density}
\usage{
plot_density(df, n = 50, ...)
}
\arguments{
\item{df}{A data frame.}

\item{n}{Threshold for max number of plots.  Defaults to 50.}

\item{...}{Optional arguments passed to geom_density().}
}
\description{
Plot a density distribution for all numeric variables.
}
\details{
The argument n sets the max number of plots allowed by the plotting function.
This is done to prevent producing a large number of plots by accident.
Change this value to override the default of 50.
}
\examples{
plot_density(iris)
}
