% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot Scatter}
\usage{
plot_scatter(df, plot_reduce = TRUE, n = 50, ...)
}
\arguments{
\item{df}{A data frame.}

\item{plot_reduce}{Boolean.  Deduplicate variable list before plotting?  Defaults to TRUE.  See Details below for more information.}

\item{n}{Threshold for number of plots passed to plot_check().  Defaults to 50.}

\item{...}{Optional arguments passed to geom_point().}
}
\description{
Scatter plots for all combinations of numeric variables.
}
\details{
The plot_reduce argument serves to deduplicate the variables plotted by plot_scatter().  For instance, when set to TRUE (default), plot_scatter() will return the (x, y) plot, but not (y, x).  This is determined alphabetically.  In no case will the (x, x) plot be produced.
}
\examples{
plot_scatter(iris)

plot_scatter(iris, plot_reduce = TRUE, n = 50, ggplot2::aes(color = Species))
}
