% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot Bar}
\usage{
plot_bar(df, n = 50)
}
\arguments{
\item{df}{A data frame.}

\item{n}{Threshold for number of plots passed to plot_check().  Defaults to 50.}
}
\value{

}
\description{
Plot a distribution for all categorical variables.
}
\examples{
plot_bar(iris)
}
