% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_check.R
\name{plot_check}
\alias{plot_check}
\title{Plot Check}
\usage{
plot_check(plot_vars, n)
}
\arguments{
\item{plot_vars}{A vector of variables to be plotted.}

\item{n}{Plot threshold as set in plot_*() function.}
}
\value{

}
\description{
A helper function to avoid plotting more than you'd prefer.
}
\details{
Function will fail if length(plot_vars) > n.
}
