% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot Density}
\usage{
plot_density(df, n = 50, ...)
}
\arguments{
\item{df}{A data frame.}

\item{n}{Threshold for number of plots passed to plot_check().  Defaults to 50.}

\item{...}{Optional arguments passed to geom_density().}
}
\value{

}
\description{
Plot a density distribution for all numeric variables.
}
\examples{
plot_density(iris)
}
