% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot Histogram}
\usage{
plot_histogram(df, n_bins = 30, n = 50)
}
\arguments{
\item{df}{A data frame.}

\item{n_bins}{Number of bins to pass to geom_histogram().  Defaults to 30.}

\item{n}{Threshold for number of plots passed to plot_check().  Defaults to 50.}
}
\value{

}
\description{
Plot a histogram distribution for all numeric variables.
}
\examples{
plot_histogram(iris)
}
